removed teamOverview.ts file and added this functionality into teams.ts so that it uses the existing teams route. Changed function which formats the team name in the URL for the filters.
