+++ server/routes/teams.ts
@@ -24,20 +24,7 @@ export default function routes({ serviceCatalogueService, teamsSummaryCountServi
     const teamSlug = getFormattedName(req, 'teamSlug')
     const team = await serviceCatalogueService.getTeam({ teamSlug })
     const products = team.products && team.products.length > 0 ? team.products.map(product => product) : null
-    try {
-      const teamAlertSummary = await teamsSummaryCountService.getTeamAlertSummary(teamSlug)
-      logger.info(`getTeamAlertSummary for team '${teamSlug}': ${JSON.stringify(teamAlertSummary, null, 2)}`)
 
-      const teamTrivyScanSummary = await teamsSummaryCountService.getTeamTrivyVulnerabilityCounts(products)
-      logger.info(`getTeamTrivyScanSummary for team '${teamSlug}': ${JSON.stringify(teamTrivyScanSummary, null, 2)}`)
-
-      const teamVeracodeScanSummary = await teamsSummaryCountService.getTeamVeracodeVulnerabilityCounts(products)
-      logger.info(
-        `getTeamVeracodeScanSummary for team '${teamSlug}': ${JSON.stringify(teamVeracodeScanSummary, null, 2)}`,
-      )
-    } catch (err) {
-      logger.error(`Error calling getTeamAlertSummary for team '${teamSlug}':`, err)
-    }
     const displayTeam = {
       id: team.t_id,
       name: team.name,
@@ -47,6 +34,45 @@ export default function routes({ serviceCatalogueService, teamsSummaryCountServi
     }
 
     return res.render('pages/team', { team: displayTeam })
+  })
+
+  router.get('/team-overview/:teamSlug', async (req, res) => {
+    const teamSlug = getFormattedName(req, 'teamSlug')
+    const team = await serviceCatalogueService.getTeam({ teamSlug })
+    const products = team.products.map(product => product)
+
+    try {
+      const teamAlertSummary = await teamsSummaryCountService.getTeamAlertSummary(teamSlug)
+      logger.info(`getTeamAlertSummary for team '${teamSlug}': ${JSON.stringify(teamAlertSummary, null, 2)}`)
+
+      const teamTrivyScanSummary = await teamsSummaryCountService.getTeamTrivyVulnerabilityCounts(products)
+      logger.info(`getTeamTrivyScanSummary for team '${teamSlug}': ${JSON.stringify(teamTrivyScanSummary, null, 2)}`)
+      const criticalAndHighTrivy = teamTrivyScanSummary.critical + teamTrivyScanSummary.high
+
+      const teamVeracodeScanSummary = await teamsSummaryCountService.getTeamVeracodeVulnerabilityCounts(products)
+      logger.info(
+        `getTeamVeracodeScanSummary for team '${teamSlug}': ${JSON.stringify(teamVeracodeScanSummary, null, 2)}`,
+      )
+      const veryHighAndHighVeracode = teamVeracodeScanSummary.veryHigh + teamVeracodeScanSummary.high
+
+      function formatTeamNameURL(teamName: string) {
+        return teamName.replace(/[ &()]/g, (char: string) => {
+          const replacements: Record<string, string> = { ' ': '+', '&': '%26', '(': '%28', ')': '%29' }
+          return replacements[char] || char
+        })
+      }
+
+      res.render('pages/teamOverview', {
+        teamName: team.name,
+        formatTeamNameURL: formatTeamNameURL(team.name),
+        teamSlug: team.slug,
+        teamAlertSummary,
+        criticalAndHighTrivy,
+        veryHighAndHighVeracode,
+      })
+    } catch (err) {
+      logger.error(`Error calling getTeamAlertSummary for team '${teamSlug}':`, err)
+    }
   })
 
   return router
